# specs/admin/openapi.yaml
openapi: 3.1.0
info:
  title: BWB Fiscal Admin API
  version: 0.1.10-draft
  description: >
    Contrato administrativo separado da API POS pública. Prefixo /admin/v1.
    Autenticação OIDC/JWT com RBAC (owner|admin|operator|auditor), distinta do
    Bearer credential_store POS. Runtime: fail_closed | injected (dev) |
    oidc_jwt (JWKS https, iss/aud exactos, role map; DEC-BO-003 / RM-BO-006).
    Segredos NÃO são devolvidos — zona SecAdm owner-only write-only (DEC-BO-001/002).
    0.1.1: PATCH scope-bindings (séries/timezone/ambiente metadados; RM-BO-002).
    0.1.2: ops visibility + audit list + secret-ref metadata sanitizado (RM-BO-003).
    0.1.3: matriz RBAC documentada; MFA adiado até IdP; sem ACL secret.reveal (RM-BO-004).
    0.1.4: SecAdm HTTP Put/Rotate/Revoke owner-only; respostas só metadados (RM-SECADM-003).
    0.1.5: listagens cadastro + PATCH status taxpayer/establishment (RM-BO-005).
    0.1.6: fundação OIDC/JWT provider-neutral documentada (RM-BO-006).
    0.1.7: observabilidade admin — health/ready/metrics + request_id (RM-BO-007).
    0.1.8: AuthorityProfile metadados HML/PRD (RM-AGTPREP-002 / DEC-BO-004); sem segredos; external_verified=false.
    0.1.9: UI SSR owner-only /admin/ui/authority-profiles (RM-AGTPREP-003); contrato HTTP inalterado; sem plaintext.
    0.1.10: SecAdm material multipart PEM/PKCS#12/credencial (RM-AGTPREP-004); password efémera; limites; ≠ AGT.
    Não pronto para produção. ≠ credenciais AGT. ≠ fornecedor IdP fixo.
  license:
...
servers:
...
security:
  - adminBearer: []
paths:
...
  /secadm/material:
    post:
      operationId: secadmPutMaterial
      summary: Importa material PEM/PKCS#12/credencial (multipart; owner)
      description: >
        Multipart: kind, environment, subject_id, name, encoding (pem|pkcs12|credential),
        material (ficheiro), password (só PKCS#12; efémera, nunca persistida/devolvida/auditada),
        expires_at opcional. Limites de tamanho fail-closed. Resposta só metadados + format_note.
        ≠ AGT real. UI: /admin/ui/secadm/material.
      tags: [SecAdm]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [kind, environment, subject_id, name, encoding, material]
              properties:
                kind:
                  type: string
                  enum: [producer_credential, producer_key, taxpayer_key, certificate]
                environment:
                  type: string
                  enum: [homologation, production]
                subject_id:
                  type: string
                name:
                  type: string
                encoding:
                  type: string
                  enum: [credential, pem, pkcs12]
                material:
                  type: string
                  format: binary
                password:
                  type: string
                  format: password
                  description: Efémera — nunca persistida
                expires_at:
                  type: string
                  format: date-time
      responses:
        "200":
          description: Metadados após import
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecAdmMaterialMetadata"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Unprocessable"
  /secadm/material/rotate:
    post:
      operationId: secadmRotateMaterial
      summary: Rotaciona material PEM/PKCS#12/credencial (multipart; owner)
      tags: [SecAdm]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [kind, environment, subject_id, name, encoding, material]
              properties:
                kind:
                  type: string
                environment:
                  type: string
                  enum: [homologation, production]
                subject_id:
                  type: string
                name:
                  type: string
                encoding:
                  type: string
                  enum: [credential, pem, pkcs12]
                material:
                  type: string
                  format: binary
                password:
                  type: string
                  format: password
                expires_at:
                  type: string
                  format: date-time
      responses:
        "200":
          description: Metadados após rotate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecAdmMaterialMetadata"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Unprocessable"
...
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        OIDC/JWT de operador (RBAC: owner|admin|operator|auditor). Distinto do
        bearerAuth POS. Só owner acede SecAdm (e subject allowlist no runtime).
        Modos: fail_closed; injected (só development); oidc_jwt com JWKS https
        e mapa de grupos configurável (DEC-BO-003). Sessão browser ≠ este Bearer
        (RM-UI-005). Nunca enviar refresh tokens neste header.
...
  schemas:
...
    SecAdmMaterialMetadata:
      allOf:
        - $ref: "#/components/schemas/SecretRefMetadata"
        - type: object
          additionalProperties: false
          required: [encoding, format_note]
          properties:
            encoding:
              type: string
              enum: [credential, pem, pkcs12]
            format_note:
              type: string
              description: Nota sanitizada (ex. pem_private_key); nunca plaintext
    SecretRefMetadata:
      type: object
      additionalProperties: false
      required: [kind, environment, subject_id, name, status, version]
      properties:
        kind:
          type: string
        environment:
          type: string
        subject_id:
          type: string
        name:
          type: string
        status:
          type: string
        fingerprint:
          type: string
        version:
          type: integer
        expires_at:
          type: string
          format: date-time
        last_verified_at:
          type: string
          format: date-time
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, request_id]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        request_id:
          type: string
  responses:
    Unauthorized:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Autenticado sem papel suficiente
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Unprocessable:
      description: Validação falhou
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"